+++ trunk/AudioSourceProtocol.h
@@ -12,6 +12,6 @@
 
 @protocol AudioSource
 
-- (AUDIO_SHORTS_PTR) getAudio:(int) msec;
+- (AUDIO_SHORTS_PTR) getAudio:(float) msec;
 
 @end
